requirements.txt
fastapi
uvicorn
mlflow
scikit-learn
pandas
pyyaml